logic/requirements.txt
psycopg2-binary
biopython
python-dotenv
fastapi
uvicorn
requests
pydantic
scikit-learn